requirements.txt
onnxruntime-gpu
opencv-python-headless
typer
tqdm
Pillow
streamlit
pandas
